+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#2004](https://github.com/teambit/bit/issues/2004) ask for approval before exporting a component to another scope (fork)
 - [#1956](https://github.com/teambit/bit/issues/1956) introduce a new flag `--codemod` for `bit export` to replace the import/require statements in the source to the newly exported scope
 - fix "Converting circular structure to JSON" error when logging a circular metadata object
 - fix exporting to a different scope than workspace configuration of `defaultScope`